Move the legacy chain check to the `check` module
And move `populated_state` to the `arbitrary` module.
